refactor: update the ebook module structure
