+++ rslang/src/app/ebook/constants/config-ebook.ts
