+++ rslang/src/app/ebook/ebook-routing.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { EbookHome } from './components/ebook-home/ebook-home.component';
-import { WordsCollections } from './components/collections/words-collections.component';
+
+import { EbookHome } from './pages/ebook-home/ebook-home.component';
+import { WordsCollections } from './components/words-collections/words-collections.component';
 import { WordsList } from './components/list-words/words-list.component';
 
 const routes: Routes = [
