+++ rslang/src/app/ebook/ebook.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
-
 import { EbookRoutingModule } from './ebook-routing.module';
-import { EbookHome } from './components/ebook-home/ebook-home.component';
-import { WordsCollections } from './components/collections/words-collections.component';
+
+import { EbookHome } from './pages/ebook-home/ebook-home.component';
+import { WordsCollections } from './components/words-collections/words-collections.component';
+
 @NgModule({
   declarations: [EbookHome, WordsCollections],
   imports: [SharedModule, EbookRoutingModule],
+  exports: [WordsCollections],
 })
 export class EbookModule {}
